fix: filter allow email domain manage course team
Allow to add course team users that aren't allowed to enroll by the filter allow email domain.
Use the normal procedure to skip the email domain,
then the manage course team user case won't block adding the user to the course team.
The filter only validates the email domain if the user is not already enrolled in the course.

fccn/nau-technical#397